Continue with process structure, add logger

+++ threadbank.c
@@ -4,27 +4,29 @@
  // Signal handler
 void sig_handler(int signo) {
     if (signo == SIGINT){// CTRL+C to abort the program
-        //if (pid_p == getpid()) { // Parent process
-        printf("\nControl+C pressed, exiting the program.\n");
-        free(fd1); free(fd2); free(request); // Free the allocated variables
-        //exec("rm *.bank"); // run rm *.bank
-        if (munmap(queue_arr, sizeof(int)*n) == -1) { // Remove mapped array
-            perror("munmap of queuer_arr failed with error: ");
+        if (pid_p == getpid()) { // Parent process
+            printf("\nControl+C pressed, exiting the program.\n");
+            free(fd1); free(fd2); free(request); // Free the allocated variables
+            //exec("rm *.bank"); // run rm *.bank
+            if (munmap(queue_arr, sizeof(int)*n) == -1) { // Remove mapped array
+                perror("munmap of queuer_arr failed with error: ");
+            }
+            if (munmap(flag, sizeof(int)*n) == -1) { // Remove mapped array
+                perror("munmap flag failed with error: ");
+            }
+            char cmd1[10] = {"rm *.bank"}; // Define command to remove all bank notes
+            char cmd2[24] = {"rm transaction_log.txt"}; // Define command to remove transaction log
+            system(cmd1); // Remove all bank notes
+            system(cmd2); // Remove transaction log
+            kill(0,SIGTERM); // Terminate main program
+            exit(EXIT_SUCCESS);
         }
-        if (munmap(flag, sizeof(int)*n) == -1) { // Remove mapped array
-            perror("munmap flag failed with error: ");
-        }
-        char cmd[10] = {"rm *.bank"}; // Define command to remove all bank notes
-        system(cmd); // Remove all bank notes
-        kill(0,SIGTERM); // Terminate main program
-        exit(EXIT_SUCCESS);
-        /*}
         else {
             fprintf(stdout, "\nDesk # %d closing.\n", i+1);
             exit(EXIT_SUCCESS);
-        }*/
+        }
     }
-    else if (signo == SIGINFO) { // && pid_p == getpid()) { // CTRL+T, master desk signal
+    else if (signo == SIGINFO && pid_p == getpid()) { // CTRL+T, master desk signal
         *flag = 1; // Raise the global flag
         printf("CTRL+T pressed, waiting for all desks to report to the master.\n");
         for (int i = 0; i < n; i++) {
@@ -42,16 +44,14 @@ void sig_handler(int signo) {
 }
 
 void master_desk(void) {
-    int s; // Shortest line help variable
+    int s; // Shortest line helping variable
 
     while(1){ //signal(SIGINT, sig_handler) != SIG_ERR || signal(SIGINFO, sig_handler) != SIG_ERR) {
         if (fgets(request, INPUT_SIZE, stdin) != NULL && strlen(request) > 1) {
-            //printf("master desk - strlen request %lu\n", strlen(request));
-            //request[strlen(request) - 1] = '\0'; // Convert last characther to NUL byte
-            //printf("+++%s+++\n", request);
+            request[strlen(request) - 1] = '\0'; // Convert last characther to NUL byte
             s = shortestline(); // Get the desk with shortest line
             queue_arr[s]++; // Increase the queue of desk s by one
-            write(fd1[2*s+WRITE], request, strlen(request)); // Write request to pipe
+            write(fd1[2*s+WRITE], request, strlen(request)+1); // Write request to pipe, include NUL byte
         }
     }
 }
@@ -72,7 +72,7 @@ int main(int argc, char *argv[]) {
     n = atoi(argv[1]);
     if (n <= 0) { printf("Invalid input. Exiting.\n"); exit(1); }
     printf("Welcome to ThreadBank manager!\n");
-    //pid_t pid_c = 0; // PID child
+    pid_t pid_c = 0; // PID child
 
     queue_arr = (int *)mmap(NULL, sizeof(int)*n, PROT_READ | PROT_WRITE, MAP_SHARED | MAP_ANONYMOUS, -1, 0);
     if (queue_arr == MAP_FAILED) {
@@ -91,40 +91,32 @@ int main(int argc, char *argv[]) {
     fd1 = malloc(2*n*sizeof(fd1)); // Master to desk, d1[0] to read and fd1[1] to write
     fd2 = malloc(2*n*sizeof(fd2)); // Desk to master
 
-    pthread_t thread_id;
-
     for (i = 0; i < n; i++) { // Create all processes and open pipes
         ((int*)queue_arr)[i] = 0; // Is it needed as calloc initializes as zero?
         if (pipe(fd1 + 2*i) == -1) { fprintf(stderr, "Pipe 1 Failed\n"); } // Open pipe fd1
         if (pipe(fd2 + 2*i) == -1) { fprintf(stderr, "Pipe 2 Failed\n"); } // Open pipe fd2
         if (fcntl(*fd1, F_SETFL, O_NONBLOCK) == -1) { fprintf(stderr, "Setting fd1 to non-blocking failed\n"); } // Set fd1 to non-blocking
         if (fcntl(*fd2, F_SETFL, O_NONBLOCK) == -1) { fprintf(stderr, "Setting fd2 to non-blocking failed\n"); } // Set fd2 to non-blocking
-        //pid_p = getpid(); // Get parent PID
-        //pid_c = fork(); // Fork process
-        //if (pid_c < 0) { printf("Fork failed.\n"); }
-        //else if (pid_c == 0) { // Child process
-        int *j = malloc(sizeof(int)); // Remember to free
-        j = &i;
-        if (pthread_create(&thread_id, NULL, desk, j) != 0) {
-            perror("Thread creation error");
+        pid_p = getpid(); // Get parent PID
+        pid_c = fork(); // Fork process
+        if (pid_c < 0) { printf("Fork failed.\n"); }
+        else if (pid_c == 0) { // Child process
+            //close(fd1[2*i+WRITE]); // Close writing end of fd1
+            //close(fd2[2*i+READ]); // Close reading end of fd2
+            desk(i, fd1, fd2, flag); // Child becomes the counter - passing the number of desk and pipes
         }
-        //close(fd1[2*i+WRITE]); // Close writing end of fd1
-        //close(fd2[2*i+READ]); // Close reading end of fd2
-        //close(fd1[2*i+READ]); // Close reading end of fd1
-        //close(fd2[2*i+WRITE]); // Close writing end of fd2
-            //desk(i, fd1, fd2, flag); // Child becomes the counter - passing the number of desk and pipes
-        //else { // Parent process
-        //    sigaction(SIGINFO,  &act, NULL); // Initialized SIGINFO handler on the master thread for requesting deposits/withdrawals
-        //    close(fd1[2*i+READ]); // Close reading end of fd1
-        //    close(fd2[2*i+WRITE]); // Close writing end of fd2
-        //}
+        else { // Parent process
+            sigaction(SIGINFO,  &act, NULL); // Initialized SIGINFO handler on the master thread for requesting deposits/withdrawals
+            close(fd1[2*i+READ]); // Close reading end of fd1
+            close(fd2[2*i+WRITE]); // Close writing end of fd2
+        }
     }
     (n == 1) ? printf("1 desk is serving.\n") : printf("%d desks are serving.\n", n);
     printf("The bank offers four different services:\nbalance (l), withdraw (w), transfer (t), deposit (d).\n");
 
-
     request = malloc(INPUT_SIZE); // Initialize request
     if (request == NULL) { printf("No memory.\n"); }
+
     /*
     int s;
